Continuation of #6256
