+++ tests/queries/0_stateless/01199_url_functions_path_without_schema_yiurule.reference
@@ -1,0 +1,2 @@
+/a/b/c
+/?query=hello world+foo+bar
