+++ tests/queries/0_stateless/01199_url_functions_path_without_schema_yiurule.sql
@@ -1,0 +1,2 @@
+SELECT path('www.example.com:443/a/b/c') AS Path;
+SELECT decodeURLComponent(materialize(pathFull('www.example.com/?query=hello%20world+foo%2Bbar'))) AS Path;
